+++ doc/whats_new.rst
@@ -153,6 +153,11 @@ Enhancements
    - DBSCAN now supports sparse input and sample weights, and should be
      faster in general. By `Joel Nothman`_.
 
+   - Add ``class_weight`` parameter to automatically weight samples by class
+     frequency for :class:`ensemble.RandomForestClassifier`,
+     :class:`tree.DecisionTreeClassifier`, :class:`ensemble.ExtraTreesClassifier`
+     and :class:`tree.ExtraTreeClassifier`. By `Trevor Stephens`_.
+
 Documentation improvements
 ..........................
 
@@ -3183,3 +3188,5 @@ David Huard, Dave Morrill, Ed Schofield, Travis Oliphant, Pearu Peterson.
 .. _Matteo Visconti di Oleggio Castello: http://www.mvdoc.me
 
 .. _Raghav R V: https://github.com/ragv
+
+.. _Trevor Stephens: http://trevorstephens.com/